HH-218702 move DatabaseSwitcher to nab-jdbc-common
