+++ nab-data-source/src/main/java/ru/hh/nab/datasource/DataSourceContextUnsafe.java
@@ -5,6 +5,7 @@ import static java.util.Optional.ofNullable;
 import java.util.function.Supplier;
 import ru.hh.nab.common.mdc.MDC;
 import ru.hh.nab.jdbc.common.DataSourceType;
+import ru.hh.nab.jdbc.common.DatabaseSwitcher;
 
 public final class DataSourceContextUnsafe {
   public static final String MDC_KEY = "db";
