+++ nab-testbase/src/main/java/ru/hh/nab/testbase/postgres/embedded/EmbeddedPostgresDataSourceFactory.java
@@ -12,9 +12,9 @@ import ru.hh.nab.datasource.DataSourceFactory;
 import static ru.hh.nab.datasource.DataSourceSettings.JDBC_URL;
 import static ru.hh.nab.datasource.DataSourceSettings.PASSWORD;
 import static ru.hh.nab.datasource.DataSourceSettings.USER;
-import ru.hh.nab.datasource.DatabaseSwitcher;
 import ru.hh.nab.datasource.healthcheck.HealthCheckHikariDataSourceFactory;
 import ru.hh.nab.datasource.monitoring.NabMetricsTrackerFactoryProvider;
+import ru.hh.nab.jdbc.common.DatabaseSwitcher;
 
 public class EmbeddedPostgresDataSourceFactory extends DataSourceFactory {
   public static final String DEFAULT_JDBC_URL = "jdbc:postgresql://${host}:${port}/${database}";
